GridAI now move to seperate Componet and System folder
