+++ src/client/main.cpp
@@ -24,7 +24,6 @@
 #include <tempo/song.hpp>
 #include <tempo/time.hpp>
 #include <tempo/entity/EntityCreationClient.hpp>
-#include <tempo/entity/SystemGridAi.hpp>
 #include <tempo/entity/Health.hpp>
 #include <tempo/entity/LevelManager.hpp>
 #include <tempo/entity/LevelRenderer.hpp>
@@ -34,6 +33,7 @@
 #include <tempo/entity/RenderHealth.hpp>
 #include <tempo/entity/Transform.hpp>
 #include <tempo/network/client.hpp>
+#include <tempo/system/SystemGridAi.hpp>
 
 #include <SFML/Audio.hpp>
 #include <SFML/System/Clock.hpp>
