+++ src/lib-tempo/include/tempo/entity/EntityCreationServer.hpp
@@ -3,10 +3,10 @@
 
 #include <anax/World.hpp>
 
+#include <tempo/component/ComponentGridAi.hpp>
 #include <tempo/entity/EntityCreation.hpp>
 #include <tempo/entity/ID.hpp>
 #include <tempo/entity/LevelManager.hpp>
-#include <tempo/entity/ComponentGridAi.hpp>
 #include <tempo/entity/PlayerRemoteServer.hpp>
 
 #include <string.h>
